backend/form_builder_project/requirements.txt
Django==6.0.4
djangorestframework
drf-spectacular
django-celery-beat
djangorestframework-simplejwt